Make spawn-daemon only defined on linux
