+++ command/spawn_daemon_universal.go
@@ -1,0 +1,9 @@
+// +build !linux
+
+package command
+
+import "errors"
+
+func (c *SpawnDaemonCommand) Run(args []string) int {
+	return c.outputStartStatus(errors.New("spawn-daemon not supported"), 1)
+}
